ai_code_error_analyzer/requirements.txt
Django==5.1.6
channels==4.1.0
daphne==4.1.2
reportlab==4.4.3
requests>=2.31.0
python-dotenv>=1.0.1
